backend/requirements.txt
# ============================================
# backend/requirements.txt
# ============================================
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
cryptography==41.0.7
websocket-client==1.6.4
email-validator==2.1.0.post1

